Refactor: move Authentication Form components into `login` page.
- The forms are exclusively used in the login page, no need to be shared.
